backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
aiohttp==3.9.0
transformers==4.35.0
torch>=2.2.0
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.5.0
beautifulsoup4==4.12.2
lxml==4.9.3
youtube-search-python==1.6.6
duckduckgo-search==3.9.6
python-dotenv==1.0.0
huggingface-hub==0.17.3
selenium==4.15.2
requests-html==0.10.0
fake-useragent==1.4.0